blockchain_PROJ/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.0
opencv-python>=4.5.0
numpy>=1.21.0
web3>=6.0.0
python-dotenv>=0.19.0
streamlit>=1.25.0
plotly>=5.15.0